# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daybook digital planner platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /editions:
    get:
      operationId: listEditions
      tags: [editions]
      summary: List editions
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, live, all]
      responses:
        "200":
          description: Editions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Edition"
    post:
      operationId: createEdition
      tags: [editions]
      summary: Create edition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edition"
  /editions/{id}:
    get:
      operationId: getEdition
      tags: [editions]
      summary: Get edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Edition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edition"
    patch:
      operationId: updateEdition
      tags: [editions]
      summary: Update edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edition"
    delete:
      operationId: deleteEdition
      tags: [editions]
      summary: Delete edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /editions/{id}/publish:
    post:
      operationId: publishEdition
      tags: [editions]
      summary: Publish edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edition"
  /editions/{id}/unpublish:
    post:
      operationId: unpublishEdition
      tags: [editions]
      summary: Unpublish edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edition"
  /editions/{id}/art:
    get:
      operationId: getEditionArt
      tags: [editions]
      summary: List per-page art files for an edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Art files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtFile"
    put:
      operationId: updateEditionArt
      tags: [editions]
      summary: Set per-page art file ids for an edition
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditionArtUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtFile"
  /editions/ai-draft:
    post:
      operationId: aiDraftEdition
      tags: [editions]
      summary: AI-draft section list and price ranges for an edition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiDraftEditionInput"
      responses:
        "200":
          description: AI draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftEditionResult"

  # ── PLANS ────────────────────────────────────────────────────────────────────
...
  /generation:
    post:
      operationId: generatePlanner
      tags: [generation]
      summary: Generate a PDF planner from a config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationInput"
      responses:
        "202":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
  /generation/{jobId}:
    get:
      operationId: getGenerationJob
      tags: [generation]
      summary: Poll generation job status
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
...
components:
  schemas:
...
    CatalogStatus:
      type: string
      enum: [draft, live]

    # ── USER ──────────────────────────────────────────────────────────────────
...
    Edition:
      type: object
      required: [id, name, slug, tier, status, createdAt]
      properties:
        id:             { type: integer }
        name:           { type: string }
        slug:           { type: string }
        description:    { type: ["string", "null"] }
        status:         { $ref: "#/components/schemas/CatalogStatus" }
        tier:
          type: string
          enum: [basic, advanced]
        themeId:        { type: ["integer", "null"] }
        stickerPackIds: { type: array, items: { type: integer } }
        insertIds:      { type: array, items: { type: integer } }
        productIds:     { type: array, items: { type: integer } }
        planIds:        { type: array, items: { type: integer } }
        oneTimePrice:   { type: ["number", "null"] }
        yearlyPrice:    { type: ["number", "null"] }
        lifetimePrice:  { type: ["number", "null"] }
        stripeProductId: { type: ["string", "null"] }
        previewImageUrl: { type: ["string", "null"] }
        perPageArtFileIds: { type: object, nullable: true, additionalProperties: true }
        createdAt:      { type: string, format: date-time }
        updatedAt:      { type: string, format: date-time }
    EditionInput:
      type: object
      required: [name, tier]
      properties:
        name:           { type: string, minLength: 1 }
        slug:           { type: string }
        description:    { type: string }
        tier:
          type: string
          enum: [basic, advanced]
        themeId:        { type: integer }
        stickerPackIds: { type: array, items: { type: integer } }
        insertIds:      { type: array, items: { type: integer } }
        productIds:     { type: array, items: { type: integer } }
        planIds:        { type: array, items: { type: integer } }
        oneTimePrice:   { type: number }
        yearlyPrice:    { type: number }
        lifetimePrice:  { type: number }
        previewImageUrl: { type: string }
    EditionUpdate:
      type: object
      properties:
        name:           { type: string, minLength: 1 }
        description:    { type: string }
        tier:
          type: string
          enum: [basic, advanced]
        themeId:        { type: ["integer", "null"] }
        stickerPackIds: { type: array, items: { type: integer } }
        insertIds:      { type: array, items: { type: integer } }
        productIds:     { type: array, items: { type: integer } }
        planIds:        { type: array, items: { type: integer } }
        oneTimePrice:   { type: ["number", "null"] }
        yearlyPrice:    { type: ["number", "null"] }
        lifetimePrice:  { type: ["number", "null"] }
        previewImageUrl: { type: string }
        perPageArtFileIds: { type: object, additionalProperties: true }
    ArtFile:
      type: object
      required: [pageId, driveFileId]
      properties:
        pageId:      { type: string }
        driveFileId: { type: string }
        filename:    { type: ["string", "null"] }
        source:      { type: ["string", "null"] }
    EditionArtUpdate:
      type: object
      required: [artFiles]
      properties:
        artFiles:
          type: array
          items:
            $ref: "#/components/schemas/ArtFile"
    AiDraftEditionInput:
      type: object
      required: [concept]
      properties:
        concept:  { type: string }
        audience: { type: string }
        tier:
          type: string
          enum: [basic, advanced]
    AiDraftEditionResult:
      type: object
      required: [name, description, suggestedSections, priceRange]
      properties:
        name:              { type: string }
        description:       { type: string }
        suggestedSections: { type: array, items: { type: string } }
        priceRange:
          type: object
          required: [oneTime, yearly, lifetime]
          properties:
            oneTime:  { type: number }
            yearly:   { type: number }
            lifetime: { type: number }
        marketingCopy: { type: string }

    # ── PLAN ──────────────────────────────────────────────────────────────────
...
    GenerationInput:
      type: object
      required: [configId]
      properties:
        configId: { type: integer }
    GenerationJob:
      type: object
      required: [jobId, status, createdAt]
      properties:
        jobId:        { type: string }
        userId:       { type: integer }
        configId:     { type: integer }
        status:
          type: string
          enum: [pending, processing, complete, failed]
        pdfUrl:       { type: ["string", "null"] }
        configJsonUrl: { type: ["string", "null"] }
        driveFileId:  { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        pageCount:    { type: ["integer", "null"] }
        createdAt:    { type: string, format: date-time }
        completedAt:  { type: ["string", "null"], format: date-time }

    # ── GOOGLE SYNC ──────────────────────────────────────────────────────────
...